+++ lib/ref_inspector.ex
@@ -3,6 +3,8 @@ defmodule RefInspector do
   Ref Inspector - Referer parser library
   """
 
+  require Logger
+
   @doc """
   Parses a referer.
   """
@@ -12,6 +14,16 @@ defmodule RefInspector do
   @doc """
   Reloads all databases.
   """
+  @spec reload() :: :ok
+  defdelegate reload(), to: RefInspector.Database
+
+  @doc false
   @spec reload_databases() :: :ok
-  defdelegate reload_databases(), to: RefInspector.Database
+  def reload_databases() do
+    Logger.info "RefInspector.reload_databases/0 has been renamed" <>
+                " to RefInspector.reload/0. The current alias will be" <>
+                " removed in a future release."
+
+    reload()
+  end
 end